I would like to change the info in side panel
for "Desk Calendar" change to "12,025 Human Era Calendar".

Calendar made by Kurzgesagt – In a Nutshell (add a hyperlink to https://www.youtube.com/@kurzgesagt).

Rephrase the following to explain the 12025, this is the website description, so I want something similar:

Delve into humanity's fascinating history and be inspired by our ancestors' achievements. We always add 10,000 years to our calendar's date to honour the collective story of humankind.

Also add paragraph about the theme of this year, again rephrase the original:

This time it's a celebration of the great journey that took our ancestors from out of Africa to the furthest reaches of the Earth. Discover an epic new story every month while keeping track of your own adventures.