Paper Weight
Made by Cristais Cá d'Oro (in bold), it was a gift when I visited a friend of my mine in Poços de Caldas, MG, Brazil.

(add something about Murano Glass, which is the style of Cristais Cá d'Oro)

It is an apple, as reminder of the anecdote of how Isaac Newton discovered the laws of gravity.